requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=0.10.0=pyhd8ed1ab_1
aiohttp=3.7.4.post0=py39h5161555_1
aom=3.3.0=h07bb92c_1
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
blinker=1.4=py_1
box2d-py=2.3.8=pypi_0
brotlipy=0.7.0=py39h5161555_1003
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2021.10.8=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py39h2804cbe_1
cffi=1.15.0=py39h52b1de0_0
chardet=4.0.0=py39h2804cbe_2
charset-normalizer=2.0.0=pyhd8ed1ab_0
clang=5.0=pypi_0
click=8.0.3=py39h2804cbe_1
cloudpickle=2.0.0=pyhd8ed1ab_0
cryptography=35.0.0=py39hfb8cd70_2
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
ffmpeg=4.4.1=h2f96316_2
flatbuffers=1.12=pypi_0
freetype=2.10.4=h17b34a0_1
future=0.18.2=py39h2804cbe_4
gast=0.4.0=pyh9f0ad1d_0
gettext=0.19.8.1=h049c9fb_1008
gmp=6.2.1=h9f76cd9_0
gnutls=3.6.13=h706517b_1
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.41.1=py39h9e1b6db_1
gym=0.21.0=py39h2804cbe_0
h5py=3.1.0=nompi_py39h99babb8_100
hdf5=1.10.6=nompi_h0fc092c_1114
icu=69.1=hbdafb3b_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.1=py39h2804cbe_1
jbig=2.1=h3422bc3_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h27ca646_0
keras=2.6.0=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py39h4d2d688_1
krb5=1.19.2=hd92b7a7_3
lame=3.100=h27ca646_1001
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=12_osxarm64_openblas
libcblas=3.9.0=12_osxarm64_openblas
libclang=11.1.0=default_h0fdd720_1
libcurl=7.79.1=h8fe1914_1
libcxx=12.0.1=h168391b_0
libdeflate=1.8=h3422bc3_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=hbdafb3b_4
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.16=h642e427_0
liblapack=3.9.0=12_osxarm64_openblas
libllvm11=11.1.0=h93073aa_2
libnghttp2=1.43.0=he4cd7f6_1
libopenblas=0.3.18=openmp_h5dd58f0_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.19.1=hccf11d3_0
libssh2=1.10.0=hb80f160_2
libtiff=4.3.0=h74060c4_2
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.1=h3422bc3_0
libxml2=2.9.12=hedbfbf4_1
libzlib=1.2.11=hee7b306_1013
llvm-openmp=12.0.1=hf3c4609_1
lz4-c=1.9.3=hbdafb3b_1
markdown=3.3.4=pyhd8ed1ab_0
matplotlib=3.4.3=py39hdf13c20_1
matplotlib-base=3.4.3=py39h5aa4fe7_1
multidict=5.2.0=py39h5161555_1
ncurses=6.2=h9aa5885_4
nettle=3.6=hc6a1b29_0
numpy=1.19.5=py39h1f3b974_2
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=habe5f53_0
openjpeg=2.4.0=h062765e_1
openssl=1.1.1l=h3422bc3_0
opt_einsum=3.3.0=pyhd8ed1ab_1
pandas=1.3.4=py39h7f752ed_1
pcre=8.45=hbdafb3b_0
pillow=8.3.2=py39ha74c66e_0
pip=21.3.1=pyhd8ed1ab_0
protobuf=3.19.1=py39hfb83b0d_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyglet=1.5.16=py39h2804cbe_1
pyjwt=2.3.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd8ed1ab_0
pysocks=1.7.1=py39h2804cbe_4
python=3.9.7=h54d631c_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.1=hedafd6a_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-learn=1.0.1=py39h12ba089_1
scipy=1.7.1=py39h5060c3b_0
setuptools=58.5.3=py39h2804cbe_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.36.0=h72a2b83_2
svt-av1=0.9.1=h07bb92c_0
swig=4.0.2=hc25a523_2
tensorboard=2.6.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py39hfb8cd70_1
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow-deps=2.6.0=0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-macos=2.6.0=pypi_0
tensorflow-metal=0.2.0=pypi_0
termcolor=1.1.0=py_2
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=he1e0b03_1
tornado=6.1=py39h5161555_2
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
tzdata=2021e=he74cb21_0
urllib3=1.26.7=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.35.1=pyh9f0ad1d_0
wrapt=1.12.1=py39h5161555_3
x264=1!161.3030=h3422bc3_1
x265=3.5=h666519e_1
xz=5.2.5=h642e427_1
yarl=1.7.2=py39h5161555_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=hee7b306_1013
zstd=1.5.0=h861e0a7_0
